Update code through chat
