+++ Physical_oM/Interface/IOpening.cs
@@ -27,10 +27,11 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Geometry;
+using BH.oM.Base;
 
 namespace BH.oM.Physical.Interface
 {
-    public interface IOpening
+    public interface IOpening : IBHoMObject
     {
         /***************************************************/
         /**** Properties                                ****/
